+++ apps/cowswap-frontend/src/modules/tokensList/pure/SelectTokenModal/MobileChainSelector.tsx
@@ -7,7 +7,7 @@ import { ChevronDown } from 'react-feather'
 
 import { Field } from 'legacy/state/types'
 
-import { TradeType } from 'modules/trade/types'
+import { TradeType } from 'modules/trade'
 
 import { ChainChip } from './ChainChip'
 import * as styledEl from './mobileChainSelector.styled'