+++ libs/ens/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [2.2.0](https://github.com/cowprotocol/cowswap/compare/ens-v2.1.0...ens-v2.2.0) (2026-02-13)
+
+
+### ✨ Features
+
+* replace toLowerCase on getAddressKey ([#6866](https://github.com/cowprotocol/cowswap/issues/6866)) ([a41b23c](https://github.com/cowprotocol/cowswap/commit/a41b23c4384cdbcde9534abcdde37441121381e7))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 2.2.0
+    * @cowprotocol/common-utils bumped to 2.2.0
+    * @cowprotocol/wallet-provider bumped to 2.1.1
+
 ## [2.1.0](https://github.com/cowprotocol/cowswap/compare/ens-v2.0.1...ens-v2.1.0) (2026-02-05)